Separate UART driver implementation
